moving data_processing/ functionality to utils/data.py
